+++ packages/blocks/src/__internal__/rich-text/quill-keyboard.ts
@@ -45,6 +45,7 @@ export class KeyboardWithEvent extends Keyboard {
         suffix: suffixText,
         event: evt,
       };
+
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       const prevented = matches.some((binding: any) => {
         if (
